pulling get initial dimensions into its own file
