+++ src/state/dimension-marshal/dimension-marshal.js
@@ -2,20 +2,12 @@
 import { type Position } from 'css-box-model';
 import invariant from 'tiny-invariant';
 import createPublisher, { type Publisher } from './publisher';
-// TODO: state folder reaching into view
-import * as timings from '../../debug/timings';
+import getInitialPublish from './get-initial-publish';
 import type {
-  DraggableId,
   DroppableId,
   DroppableDescriptor,
-  DroppableDimension,
-  DraggableDimension,
   DraggableDescriptor,
-  DraggableDimensionMap,
-  DroppableDimensionMap,
-  DimensionMap,
   LiftRequest,
-  ScrollOptions,
   Critical,
 } from '../../types';
 import type {
@@ -255,67 +247,6 @@ export default (callbacks: Callbacks) => {
     entry.callbacks.scroll(change);
   };
 
-  const getInitialPublish = (
-    critical: Critical,
-    scrollOptions: ScrollOptions,
-    windowScroll: Position,
-  ): StartPublishingResult => {
-    const timingKey: string = 'Initial collection from DOM';
-    timings.start(timingKey);
-
-    const home: DroppableDescriptor = critical.droppable;
-
-    const droppables: DroppableDimensionMap = Object.keys(entries.droppables)
-      .map((id: DroppableId): DroppableEntry => entries.droppables[id])
-      // Exclude things of the wrong type
-      .filter(
-        (entry: DroppableEntry): boolean => entry.descriptor.type === home.type,
-      )
-      .map(
-        (entry: DroppableEntry): DroppableDimension =>
-          entry.callbacks.getDimensionAndWatchScroll(
-            windowScroll,
-            scrollOptions,
-          ),
-      )
-      .reduce(
-        (previous: DroppableDimensionMap, dimension: DroppableDimension) => {
-          previous[dimension.descriptor.id] = dimension;
-          return previous;
-        },
-        {},
-      );
-
-    const draggables: DraggableDimensionMap = Object.keys(entries.draggables)
-      .map((id: DraggableId): DraggableEntry => entries.draggables[id])
-      .filter(
-        (entry: DraggableEntry): boolean =>
-          entry.descriptor.type === critical.draggable.type,
-      )
-      .map(
-        (entry: DraggableEntry): DraggableDimension =>
-          entry.getDimension(windowScroll),
-      )
-      .reduce(
-        (previous: DraggableDimensionMap, dimension: DraggableDimension) => {
-          previous[dimension.descriptor.id] = dimension;
-          return previous;
-        },
-        {},
-      );
-
-    timings.finish(timingKey);
-
-    const dimensions: DimensionMap = { draggables, droppables };
-
-    const result: StartPublishingResult = {
-      dimensions,
-      critical,
-    };
-
-    return result;
-  };
-
   const stopPublishing = () => {
     // This function can be called defensively
     if (!collection) {
@@ -363,7 +294,12 @@ export default (callbacks: Callbacks) => {
       critical,
     };
 
-    return getInitialPublish(critical, request.scrollOptions, windowScroll);
+    return getInitialPublish({
+      critical,
+      windowScroll,
+      entries,
+      scrollOptions: request.scrollOptions,
+    });
   };
 
   const marshal: DimensionMarshal = {
